third_party/WebKit/LayoutTests/fast/css/empty-pseudo-class.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<style>
    div > div { color: red }
    .empty:empty, .not-empty:not(:empty) { color: green }
</style>
<div id="sandbox">
    <div class="empty"></div>
    <div class="empty"><!-- Just a comment --></div>
    <div class="not-empty"> </div>
    <div class="not-empty">
						.
					</div>
    <div class="not-empty"><blockquote></blockquote></div>
    <div id="appendParent" class="not-empty"></div>
</div>
<script>
    description(":empty selector tests");

    appendParent.appendChild(document.createElement("div"));

    var green = "rgb(0, 128, 0)";

    var testDivs = sandbox.querySelectorAll("#sandbox > div");
    for (var i = 0; i < testDivs.length; i++)
        shouldBe("getComputedStyle(testDivs["+i+"]).color", "green");
</script>
