third_party/blink/web_tests/fast/css/empty-pseudo-class-expected.txt

:empty selector tests

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(testDivs[0]).color is green
PASS getComputedStyle(testDivs[1]).color is green
PASS getComputedStyle(testDivs[2]).color is green
PASS getComputedStyle(testDivs[3]).color is green
PASS getComputedStyle(testDivs[4]).color is green
PASS getComputedStyle(testDivs[5]).color is green
PASS successfullyParsed is true

TEST COMPLETE

.
